coltrane/requirements.txt
django-tagging
markdown
-e git://github.com/mgan59/python-pinboard.git